backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
kaggle==1.7.4.2
langchain==0.3.23
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.24
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
protobuf==5.29.4
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
text-unidecode==1.3
tokenizers==0.21.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.51.0
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
zstandard==0.23.0
